hack27/pyproject.toml
[project]
name = "hack27"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.0",
    "python-dotenv>=1.0.1",
    "manim>=0.18.1",
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
    "rich>=13.7.0",
    "jsonschema>=4.19.0",
    "langchain-google-genai>=0.1.0",
    "flask>=3.1.2",
]
